Functions and types of ThermovisorData

Base.length — Method

Base.length(c::CentredObj)

Total number of values needed to create CentredObj of specified type

Main.ThermovisorData._inbounds_flag — Method

_inbounds_flag(L,D,max_length,min_length)

Unsafe version of check! number of  rows in `D` should be the same as the number of 
elements in `L`
Returns Bool flag of all row not containing NaN's and lying within the min_length to max_length range

Main.ThermovisorData.add_distrib_point! — Method

add_distrib_point!(points,distrib,point,value)

Internal fucntion to add the point to distribution

Main.ThermovisorData.along_line_distribution — Method

Function evaluates the matrix values distribution along the line specified by two coordinates, 
    img - input image 
	returns the tuple of two vectors: coordinates and values 

see [`ImageDraw.bresenham(img, y0, x0, y1, x1, color)`](@ref)

returns 
points - vector of coordinates along the line
distrib - distribution

Main.ThermovisorData.along_line_distribution_xiaolin_wu — Method

along_line_distribution_xiaolin_wu(img::AbstractMatrix{T}, y0, x0, y1, x1) where T

Evaluates the value matrix content along the line with endpoint coordinates x0,y0,y1,x1, returns indices of all points. As far as Wu's algorithm returns two adjacent points the value is evaluated as an average of two point obtained with Wu's algorithm

see xiaolin_wu function from ImageDraw

Main.ThermovisorData.along_mask_line_distribution — Function

along_mask_line_distribution(imag::AbstractMatrix,c::CentredObj,direction_angle=0.0,line_length=10.0;
                                                                                            length_per_pixel=1.0,
                                                                                            use_wu::Bool=false)

The same as `within_mask_line_points_distribution` but returns the line length along the coordinates within the 
image.

`line_length` - the length of line in the same units as `length_per_pixel`.
The calibration `using mm_per_pixel``, returns calibrated length along the line

Main.ThermovisorData.angular_distribution_statistics — Method

angular_distribution_statistics(angles,along_length_coordinate,distrib;
                            max_length=-1.0,is_use_student::Bool=true)

Function evaluates average temperature distribution vs angle of orientation

Main.ThermovisorData.area — Method

Ealuates the surface area

Main.ThermovisorData.cent_to_flag — Method

cent_to_flag(c::CentredObj,sz::Tuple{Int,Int};external=false)

Converts CentredObj to bitmatrix  of size sz

Main.ThermovisorData.cent_to_flag — Method

cent_to_flag(::Type{T},c::CentredObj,sz::Tuple{Int,Int};external=false) where T<:FlagMatrix

Converts centred obj to BitMatrix of the Matrix of bool see `FlagMatrix

Main.ThermovisorData.center — Method

center(c::CentredObj)

Returns objects central point

Main.ThermovisorData.copyobj — Method

    copyobj(c::T) where T<:CentreObj

Copies the CentreObj creating new instance

Main.ThermovisorData.diagonal_points — Method

diagonal_points(c::Union{SquareObj,CircleObj})

Returns diagonal points in row-column coordinates

Main.ThermovisorData.dimensions — Method

dimensions(c::CentredObj)

Return dimentional parameters (vector)

Main.ThermovisorData.draw! — Method

draw!(image::Matrix{Float64},c::CentredObj;fill=false,thickness::Int=55,color::RGB{Float64}=RGB{Float64}(0,1,0), kwargs...)

Draws CentreObj inside the image.

image - image
c - object 
fill - if true the interior of the object will be filled 
thickness - the thickness of the object's frame
color - frame and filling color

Main.ThermovisorData.draw — Method

draw(c::CentredObj;kwargs...)

Returns CentredObj image

Main.ThermovisorData.external_flag_from_marker — Method

external_flag_from_marker(markers::Matrix{Int};label=0,external=true)

Converts patterns markers matrix (image where each pattern marked with its own integer) to flag matrix of region external (if external is true) of internal otherwise , if label equals zero looks for pattern with maximal number of pixels

Main.ThermovisorData.fill_from_vect! — Method

fill_from_vect!(c::CentredObj, v::AbstractVector)

Fills CentreObj parameters from the vector [centerindex1,centerindex2,dimention1,dimention2,...]

Main.ThermovisorData.fill_im! — Method

`fill_im!(img,c::CentreObj)`

Fills bitmatrix `img` in a way that all pixels which are 
within the CentreObj are set to true.  See also `is_within`

Main.ThermovisorData.fill_im_external! — Method

fill_im_external!(img::FlagMatrix,c::CentredObj)

Fills image matrix img in a way that all pixels which are not within the CentreObj set to true. See also is_within

Main.ThermovisorData.fill_x0! — Method

fill_x0!(x0,im_bin::AbstractMatrix,c::CentredObj)

Fills the optimization starting vector by seraching the centre of the image `im_bin`

Main.ThermovisorData.fill_x0! — Method

fill_x0!(x0,im_bin::FlagMatrix,::CircleObj)

Fills starting vector for the optimization of CentredObj

Main.ThermovisorData.filter_image! — Method

filter_image!(imag::AbstractMatrix,flag::BitMatrix)

Returns `FilteredImage` taking all elements of imag which are not external_region_flag

Main.ThermovisorData.filter_image! — Method

filter_image!(imag::RescaledImage{Float64},external_region_flag::FlagMatrix)::FilteredImage

In-place filtering of RescaledImage

Main.ThermovisorData.filter_image — Method

filter_image(imag::AbstractMatrix,c::CentredObj;external=false)

Filters image according to centered object creating new image if external is true than as a filtering flag the inverse of centered object image is taken

Main.ThermovisorData.filter_image — Method

filter_image(imag::RescaledImage,c::CentredObj;external=false)

Filters image according

Main.ThermovisorData.filter_image — Method

    `filter_image(imag::RescaledImage,markers;label=0)`

Funtion zeroes all pixels of the image, except those belonging to the specified pattern. image - rescaled image (see RescaledImage type) markers - the matrix of the same size as the input image, each element of this matrix has unique value-label associated with some pattern. Function label_components returns the markers matrix. (optional) - the value of the label to be selected as a pattern marker

Function returns FIlteredImage type object

Main.ThermovisorData.find_temperature_files — Function

    `find_temperature_files(folder::AbstractString)`

Searchs the folder for thermal images files using is_temperature_file Returns dictionary Dict{String,Pair{Float64,String}} with keys parts of files matched using is_temperature_file, values - are temperature pairs of Float64 => full-file-name When file name contains "BB" it supposed to be the blackbody themperature distribution

Main.ThermovisorData.fit_centred_obj! — Function

fit_centred_obj!(c::CentredObj,im_bin::FlagMatrix,x0=nothing;
                            optimizer = NelderMead())

Fits centred obj to binary image by adjusting centre coordinates and dimentions
`x0` - staring vector

Main.ThermovisorData.fit_centred_obj! — Function

fit_centred_obj!(c::CentredObj,image::FilteredImage,x0=nothing;fit_reduced::Bool=true)

Fits CentredObj (modified) to filtered image (not modified) fit_reduced flag (default=true) indicates what version of the image should be fitted if true - reduced otherwise - full image

Main.ThermovisorData.full_image_flag — Method

full_image_flag(filtered_im::FilteredImage)

Returns the BItMatrix flag of filtered pattern in the whole image.
```
    im_rescale = RescaledImage()
```

Main.ThermovisorData.image_discr — Method

Calculates the scalar distance between two matrices by checking the equality of their elements

Main.ThermovisorData.image_fill_discr — Method

Function returns the function to evaluate the discrepancy between CentredObj and the matrix, this function is used during the fitting procedure

Main.ThermovisorData.is_temperature_file — Method

`is_temperature_file(file_name::AbstractString)`

Checks if the file with `file_name` has an appropriate name for thermovisor temperature distribution file

Main.ThermovisorData.is_within — Method

Function to check if indices are within the object

Main.ThermovisorData.line_points_to_along_length — Method

line_points_to_along_length(along_line_points::Vector{T},line_points) where T

Converts Cartesian indices of `along_line_points` to the length along line

Main.ThermovisorData.line_within_mask — Method

line_within_mask(c::CentredObj,ang::Float64,line_length::Int)

Function returns endpoint of the line lying fully within the mask - tuple of four point which can be directly splatted to the alonglinedistribution

ang - angle in degrees line_length - the length of line

Main.ThermovisorData.line_within_mask — Method

line_within_mask(c::CircleObj,ang,line_length)

Returns two endpoints of the line lying totally inside the CentredObj

Main.ThermovisorData.marker_image — Method

markerimage(rescaled::RescaledImage,levelthreshold::Float64,distance_threshold::Float64=1e-3)

Markers image patterns, input umage is RescaledImage image type, levelthreshold - should be between 0.0 and 1.0 distancethreshold - criterium of image binarization after distance transform

returns markers - matrix of Int's with the same size as the input matrix, each element of markers is the label index of individual patterns of the initial image

Main.ThermovisorData.obj_from_vect — Method

obj_from_vect(::Type{CentredObj},v::AbstractVector)

Creates object from parameters vector, first two arguments are center point other are dimentions [center[1],center[2],dimentions[1],...]

Main.ThermovisorData.plot_along_line_distribution — Method

plot_along_line_distribution(along_line_length,along_line_distribution;
                                    length_scaler::Float64=1.0,
                                    is_centered::Bool=true,kwargs...)

Plots temperature distribution along the line along_line_length - coordinates, along_line_distribution - values of temperature, length_scaler - length scaler (can be used to convert pixels to the actual length) is_centered - the line length is converted to the coordinates with zero value in the centre of the CentredObj

Main.ThermovisorData.plot_angular_distribution_statistics — Method

plot_angular_distribution_statistics(angles,mean_D::T,std_D::T,
            lower_bound::Union{T,Nothing}=nothing,upper_bound::Union{T,Nothing}=nothing;
            length_scaler::Float64=1.0,
            label=nothing,
            minorgrid=true,
            gridlinewidth=2,
            title="Average temperature angular distribution",framestyle = :box,
            dpi=600,xlabel = L"Angle  ,°", ylabel=L"Temperature °C",
            kwargs...)      where T<:AbstractVector

The same as plot_radial_distribution_statistics but plots averaged angular distribution

Main.ThermovisorData.plot_radial_distribution_statistics — Method

plot_radial_distribution_statistics(L,mean_D::T,std_D::T,
    lower_bound::Union{T,Nothing}=nothing,upper_bound::Union{T,Nothing}=nothing;
            length_scaler::Float64=1.0,
            is_centered::Bool=true,label=nothing,
            minorgrid=true,gridlinewidth=2,title="Average temperature radial distribution",
            framestyle = :box,
            dpi=600,xlabel = "Distance  across the sample ,mm", ylabel="Temperature  °C",
            kwargs...)      where T<:AbstractVector

Plots radial ditribution averaged value, confidence bounds and confidence bounds
multiplied by the Student's coefficient

Main.ThermovisorData.points_within_line! — Method

points_within_line!(imag::AbstractMatrix,line_points::AbstractVector)

Forces all line points to lie within the possible region according toe the image size

Main.ThermovisorData.radial_distribution — Method

radial_distribution(imag::AbstractMatrix,c::CentredObj,angles_range::AbstractRange,line_length;mm_per_pixel=1.0)

Calls `along_mask_line_distribution` on lines oriented with some angles range and puts the resulting 
distribution into one matrix j'th column of this matrix corresponds to the distribution along the line oriented with ang[j] angle

Main.ThermovisorData.radial_distribution_statistics — Method

radial_distribution_statistics(along_length_coordinate,distrib;length_per_pixel=1.0,is_use_student=true)

This function evaluates mean radial diatribution, it's standard deviation and student's coefficient Input arguments along_length, distrib - distribution matrix. All rows of distrib which contains NaNs will be droped.

Optional: max_length - maximal value of alonglengthcoordinate to be includet in to the statistics evaluation is_use_student - flag if use students's coefficient

Main.ThermovisorData.read_temperature_file — Method

read_temperature_file(f_name::AbstractString)

Reads temeprature file f_name is a full file name

Main.ThermovisorData.rearranged_diagonal — Method

rearranged_diagonal(c::Union{SquareObj,CircleObj})

Returns diagonal points in x-y coordinates

Main.ThermovisorData.student_coefficient — Method

student_coefficient(degrees_of_freedom::Int, probability; digits::Int = 3, side::Int = 2)

Evaluates Student's distribution coefficient

Main.ThermovisorData.to_rgb — Method

to_rgb(image::Matrix{Float64};color_scheme::String="")

Converts matrix to rgb martix by applyting the color scheme using applycolourmap function from PerceptualColourMaps

Main.ThermovisorData.within_mask_line_points_distribution — Function

within_mask_line_points_distribution(imag::AbstractMatrix,c::CentredObj,direction_angle=0.0,line_length=10.0;use_wu::Bool=false)

Function evaluates the distribution of values in imag matrix along the line with length line_length in pixels oriented with the angle direction_angle in degrees with respect to the posistive direction of oX (column index increase), this line lies within the mask (CentreObj) and goes through its center.

Function returns:

    `points`  - vector of `CartesianIndex` of image's points lying on the line

    `distrib` - distribution of values

    `line_points` - endpoints of line the Tupple of (left_x,left_Y,right_x,right_y)

Main.ThermovisorData.CentredObj — Type

    `CentredObj` is a sort of ROI objects. It has centre coordinates, object's center can be 
    anywhere with respect to the image indices. ROI also has one or more dimentions (in pixels)
    coordinates of centre are equal to CartesianIndices, first element is the row index, 
    the second element is the column index!! (y and x coordinate)
    This is opposite to the ImageDraw, where first Point coordinate corresponds to the column index and 
    the second one to the row index
        
To impement `CentredObj` abstraction one needs to implement:
[`is_within`](@ref) - function to check if inds are within the `CentredObj`

[`line_within_mask`](@ref) - function to check if all line points are within the `CentredObj`

[`fill_x0!`](@ref) - function to fill the optimization starting vector during `CentredObj` 
fitting the image

[`convert_to_drawavable`](@ref) fucntion to convert the `CentredObj` to a drawable obj for `ImageDraw`

Main.ThermovisorData.CircleObj — Type

Circle object with defined diemeter

Main.ThermovisorData.FilteredImage — Type

The returning type of `filter_image` function

:full - filtered rescaled image of the same size as the input with all pixels which are not the part of the pattern with label value 

:region_indices - cartesian indices of the pattern in the input image

:reduced - rectangular image with zeroes removed as much as possible  (the 	scaling of this image is the same as of the input `imag.initial` see `RescaledImage` type )

:reduced_flag - bitmatrix version of (reduced)

Main.ThermovisorData.RectangleObj — Type

Rectangular object with defined two sides

Main.ThermovisorData.RescaledImage — Type

`RescaledImage` - structure stores the image data mapped to region  [0,1]
initial data before rescaling, min max values and image after the rescaling
we need rescaled image to use greyscale conversion, distance trasform 
        initial  - initial image before rescaling
        sz - size of the image
        min - min and max values 
        max  - 
        im - image with all values from 0 to 1

Main.ThermovisorData.SquareObj — Type

Square with defined center and side