Functions and types of ThermovisorData

Core.Type — Method

(::Type{T})() where T<:CentredObj

Empty object constructor

Main.ThermovisorData.CentredObj — Type

`CentredObj` is a sort of region of interest (ROI) marker object.

CentredObj has centre coordinates, object's center can be anywhere with respect to the image indices. ROI also has one or more size parameters (in pixels) coordinates of centre are equal to CartesianIndices, first element is the row index, the second element is the column index!! (y and x coordinate) This is opposite to the ImageDraw, where first Point coordinate corresponds to the column index and the second one to the row index. CentredObj can also be used as for indexing image[c] - returns all elements of image within c, image[c]=x sets all elements of image to the values of x, x firstindex should be 1. CentredObj can also be used to set all image points within the ROI to a single value. e.g. image[c] = 30

To impement CentredObj abstraction one needs to implement:

is_within - function to check if inds are within the CentredObj

line_within_mask - function to check if all line points are within the CentredObj

fill_x0! - function to fill the optimization starting vector during CentredObj fitting the image

convert_to_drawable fucntion to convert the CentredObj to a drawable obj for ImageDraw

parnumber function which returns the number of parameters

Main.ThermovisorData.CircleObj — Type

Circle object with defined diemeter

Main.ThermovisorData.FilteredImage — Type

    Type to store image with filtered temperature region

:full - filtered rescaled image of the same size as the input with all pixels which are not the part of the pattern with label value

:region_indices - cartesian indices of the pattern in the input image

:reduced - image of reduced size where all not-inpatter pixels removed (the scaling of this image is the same as of the input imag.initial see RescaledImage type )

:reduced_flag - bitmatrix version of (reduced)

Main.ThermovisorData.RectangleObj — Type

Rectangular object with defined two sides

Main.ThermovisorData.RescaledImage — Type

    RescaledImage - structure stores the image data mapped to region  [0,1]

Fields:

initial - initial image before rescaling

sz - size of the image

min - minimum value

max - maximum value

im - image with all values from 0 to 1

Main.ThermovisorData.SquareObj — Type

Square with defined center and side

Base.getindex — Method

Base.getindex(img::AbstractMatrix,c::CentredObj)

CentredObj can be used for matrix indexing, image[centred_object] - returns the vector of temperatures of all points of image lying within the centred_object of CentredObj

Base.length — Method

Base.length(c::CentredObj)

Total number of values needed to create CentredObj of specified type

Base.setindex! — Method

Base.setindex!(img::Matrix,x::Array,c::CentredObj)

img[c]=x assignes all x elements to the elements of img with indices lying within the CentredObj c

Base.setindex! — Method

Base.setindex!(img::Matrix{T},x::Number,c::CentredObj) where T

Setting all elements within the CentredObj to a single value

Main.ThermovisorData._inbounds_flag — Method

_inbounds_flag(L,D,max_length,min_length)

Unsafe version of check! number of rows in D should be the same as the number of elements in L Returns Bool flag of all row not containing NaN's and lying within the minlength to maxlength range

Main.ThermovisorData.add_distrib_point! — Method

add_distrib_point!(points,distrib,point,value)

Internal fucntion to add the point to distribution

Main.ThermovisorData.along_line_distribution — Method

along_line_distribution(img::AbstractMatrix{T},x0,y0,x1,y1) where T

Function evaluates matrix values distribution along the line specified by two coordinates, img - input image returns the tuple of two vectors: coordinates and values see ImageDraw.bresenham for details of finding the points of the line

returns points - vector of coordinates along the line distrib - distribution

Main.ThermovisorData.along_line_distribution_xiaolin_wu — Method

along_line_distribution_xiaolin_wu(img::AbstractMatrix{T}, y0, x0, y1, x1) where T

Evaluates the value matrix content along the line with endpoint coordinates x0,y0,y1,x1, returns indices of all points. As far as Wu's algorithm returns two adjacent points the value is evaluated as an average of two point obtained with Wu's algorithm

see xiaolin_wu function from ImageDraw

Main.ThermovisorData.along_mask_line_distribution — Function

along_mask_line_distribution(imag::AbstractMatrix,c::CentredObj,direction_angle=0.0,line_length=10.0;
                                                                                            length_per_pixel=1.0,
                                                                                            use_wu::Bool=false)

The same as within_mask_line_points_distribution but returns the line length along the coordinates within the image.

line_length - the length of line in the same units as length_per_pixel. The calibration using mm_per_pixel`, returns calibrated length along the line

Main.ThermovisorData.angular_distribution_statistics — Method

angular_distribution_statistics(angles,along_length_coordinate,distrib;
                            max_length=-1.0,is_use_student::Bool=true)

Function evaluates average temperature distribution vs angle of orientation

Main.ThermovisorData.area — Method

area(c::CentredObj)

Ealuates the surface area in pixels

Main.ThermovisorData.cent_to_flag — Method

cent_to_flag(c::CentredObj,sz::Tuple{Int,Int};external=false)

Converts CentredObj to bitmatrix of size sz

Main.ThermovisorData.cent_to_flag — Method

cent_to_flag(::Type{T},c::CentredObj,sz::Tuple{Int,Int};external=false) where T<:FlagMatrix

Converts centred obj to BitMatrix of the Matrix of bool see `FlagMatrix

Main.ThermovisorData.center — Method

center(c::CentredObj)

Returns objects central point

Main.ThermovisorData.convert_to_drawable — Method

convert_to_drawable(::CentredObj)

Converts CentredObj to a drawable structure appropriate to the ImageDraw draw function, polygon,ellipse see [ImageDraw.draw] function

Main.ThermovisorData.copyobj — Method

    copyobj(c::T) where T<:CentreObj

Copies the CentredObj creating new instance

Main.ThermovisorData.count_separate_patterns — Method

count_separate_patterns(markers::Matrix{Int})

This function takes matrix of markers see marker_image and calculates the number of separate patterns

Main.ThermovisorData.diagonal_points — Method

diagonal_points(c::Union{SquareObj,CircleObj})

Returns diagonal points in row-column coordinates

Main.ThermovisorData.dimensions — Method

dimensions(c::CentredObj)

Return dimentional parameters (vector)

Main.ThermovisorData.draw! — Method

draw!(image::Matrix{Float64},c::CentredObj;fill=false,thickness::Int=55,color::RGB{Float64}=RGB{Float64}(0,1,0), kwargs...)

Draws CentreObj inside the image.

image - image

c - object

fill - if true the interior of the object will be filled

thickness - the thickness of the object's frame

color - frame and filling color

Main.ThermovisorData.draw — Method

draw(c::CentredObj;kwargs...)

Returns CentredObj image of minimal possible size

Main.ThermovisorData.external_flag_from_marker — Method

external_flag_from_marker(markers::Matrix{Int};label=0,external=true)

Converts patterns markers matrix (image where each pattern marked with its own integer) to flag matrix of region external (if external is true) of internal otherwise , if label equals zero looks for pattern with maximal number of pixels

Main.ThermovisorData.fill_from_vect! — Method

fill_from_vect!(c::CentredObj, v::AbstractVector)

Fills CentreObj parameters from the vector [centerindex1,centerindex2,dimention1,dimention2,...]

Main.ThermovisorData.fill_im! — Method

fill_im!(img,c::CentredObj)

Fills bitmatrix img in a way that all pixels which are within the CentredObj are true and false otherwise.

Main.ThermovisorData.fill_im_external! — Method

fill_im_external!(img::FlagMatrix,c::CentredObj)

Fills image matrix img in a way that all pixels which are not within the CentreObj set to true. See also is_within

Main.ThermovisorData.fill_vect! — Method

fill_vect!(x::AbstractVector, c::CentredObj)

Converts CentredObj to vector

Main.ThermovisorData.fill_x0! — Method

    fill_x0!(x0,im_bin::AbstractMatrix,c::CentredObj)

Fills the optimization starting vector by seraching the centre of the image im_bin

Main.ThermovisorData.fill_x0! — Method

fill_x0!(x0,im_bin::FlagMatrix,::CircleObj)

Fills starting vector for the optimization of CentredObj

Main.ThermovisorData.filter_image! — Method

filter_image!(imag::AbstractMatrix,flag::BitMatrix)

Returns FilteredImage taking all elements of imag which are not externalregionflag

Main.ThermovisorData.filter_image! — Method

filter_image!(imag::RescaledImage{Float64},external_region_flag::FlagMatrix)::FilteredImage

In-place filtering of RescaledImage

Main.ThermovisorData.filter_image — Method

filter_image(imag::AbstractMatrix,c::CentredObj;external=false)

Filters image according to centered object creating new image if external is true than as a filtering flag the inverse of centered object image is taken

Main.ThermovisorData.filter_image — Method

filter_image(imag::RescaledImage,c::CentredObj;external=false)

Filters image according

Main.ThermovisorData.filter_image — Method

    `filter_image(imag::RescaledImage,markers;label=0)`

Funtion zeroes all pixels of the image, except those belonging to the specified pattern. image - rescaled image (see RescaledImage type) markers - the matrix of the same size as the input image, each element of this matrix has unique value-label associated with some pattern. Function label_components returns the markers matrix. (optional) - the value of the label to be selected as a pattern marker

Function returns FilteredImage object

Main.ThermovisorData.find_temperature_files — Function

    `find_temperature_files(folder::AbstractString)`

Searchs the folder for thermal images files using is_temperature_file Returns dictionary Dict{String,Pair{Float64,String}} with keys parts of files matched using is_temperature_file, values - are temperature pairs of Float64 => full-file-name When file name contains "BB" it supposed to be the blackbody themperature distribution

Main.ThermovisorData.fit_all_patterns — Method

fit_all_patterns(img::RescaledImage,::Type{T}=CircleObj;
                                        level_threshold::Float64=0.8,
                                        distance_threshold::Float64=1e-3,
                                        max_centred_objs::Int=100,
                                        optimizer::Optim.ZerothOrderOptimizer = NelderMead(),
                                        options::Optim.Options=DEFAULT_FITTING_OPTIONS[]) where T<:CentredObj

Function fits all patterns of the image img to the vector of CentredObj ROI objects. The type of ROI should be provided as a second arguments (by default it is a CircleObj)

img - input image of RescaledImage type

For other input arguments see marker_image and fit_centred_obj!(c::CentredObj,im_bin::FlagMatrix)

Main.ThermovisorData.fit_centred_obj! — Method

fit_centred_obj!(c::CentredObj,image::FilteredImage;
                                    starting_point::Union{Nothing,Vector{Float64}}=nothing,fit_reduced::Bool=true,
                                    optimizer::Optim.ZerothOrderOptimizer = NelderMead(),options::Optim.Options=DEFAULT_FITTING_OPTIONS[])

Fits CentredObj (modified) to filtered image (not modified) fit_reduced flag (default=true) indicates what version of the image should be fitted if true - reduced otherwise - full image. For other input arguments see fit_centred_obj!(c::CentredObj,im_bin::FlagMatrix)

Main.ThermovisorData.fit_centred_obj! — Method

fit_centred_obj!(c::CentredObj,im_bin::FlagMatrix;
                            starting_point::Union{Nothing,Vector{Float64}}=nothing,
                            optimizer::Optim.ZerothOrderOptimizer = Optim.NelderMead(), 
                            options::Optim.Options=Optim.Options())

Fits CentredObj to binary image pattern (region of units) by adjusting centre coordinates and dimentions using zeroth-order optimizers from Optim.jl package.

Input variables:

c - CentredObj (modified)

im_bin - binarized image (BitMatrix or Matrix{Bool})

(optional)

startingpoint - staring vector (uses [`fillx0!`](@ref) function to fill starting point by default)

optimizer - zeroth-order optimizer from Optim.jl package

options - optimization options from Optim.jl package

Main.ThermovisorData.full_image_flag — Method

full_image_flag(filtered_im::FilteredImage)

Returns the BitMatrix flag of filtered pattern in the whole image.

Can be used as index matrix in the full image e.g.: filtered_image.full.initial[full_image_flag(filtered_image)] will return all elements which belong to the pattern

Main.ThermovisorData.image_discr — Method

image_discr(im1,im2)

Calculates the scalar distance between two matrices by checking the equality of their elements

Main.ThermovisorData.image_fill_discr — Method

image_fill_discr(image::AbstractMatrix,c::CentredObj)

Function returns the function to evaluate the discrepancy between CentredObj and the matrix, this function is used during the fitting procedure

Main.ThermovisorData.is_temperature_file — Method

    `is_temperature_file(file_name::AbstractString)`

Checks if the file with file_name has an appropriate name for thermovisor temperature distribution file

Main.ThermovisorData.is_within — Method

is_within(c::CentredObj,_)

Function to check if indices are within CentredObj

Main.ThermovisorData.is_within — Method

is_within(c::CentredObj,i::CartesianIndex)

CartesianIndex support

Main.ThermovisorData.is_within_iterator — Method

is_within_iterator(img::AbstractMatrix,c::CentredObj)

Iterator over all CartesianIndices within the img which are within the CentredObj c

Main.ThermovisorData.line_points_to_along_length — Method

line_points_to_along_length(along_line_points::Vector{T},line_points) where T

Converts Cartesian indices of `along_line_points` to the length along line

Main.ThermovisorData.line_within_mask — Method

line_within_mask(c::CentredObj,ang::Float64,line_length::Int)

Function returns endpoint of the line lying fully within the mask - tuple of four point which can be directly splatted to the alonglinedistribution

ang - angle in degrees

line_length - the length of line

Main.ThermovisorData.line_within_mask — Method

line_within_mask(c::CircleObj,ang,line_length)

Returns two endpoints of the line lying totally inside the CentredObj

Main.ThermovisorData.marker_image — Method

marker_image(rescaled::RescaledImage,level_threshold::Float64,distance_threshold::Float64=1e-3)

Markers image patterns, input umage is RescaledImage image type, levelthreshold - should be between 0.0 and 1.0 distancethreshold - criterium of image binarization after distance transform

returns markers - matrix of Int's with the same size as the input matrix, each element of markers is the label index of individual patterns of the initial image

Main.ThermovisorData.mean_within_mask — Method

mean_within_mask(img::AbstractMatrix,c::CentredObj)

Evaluates the average temperature of all points within the CentredObj marker

Main.ThermovisorData.obj_from_vect — Method

obj_from_vect(::Type{CentredObj},v::AbstractVector)

Creates object from parameters vector, first two arguments are center point other are dimentions [center[1],center[2],dimentions[1],...]

Main.ThermovisorData.parnumber — Method

parnumber(::Type{T}) where T<:CentredObj

Returns total number of parameters needed to create new object

Main.ThermovisorData.plot_along_line_distribution — Method

plot_along_line_distribution(along_line_length,along_line_distribution;
                                    length_scaler::Float64=1.0,
                                    is_centered::Bool=true,kwargs...)

Plots temperature distribution along the line along_line_length - coordinates, along_line_distribution - values of temperature, length_scaler - length scaler (can be used to convert pixels to the actual length) is_centered - the line length is converted to the coordinates with zero value in the centre of the CentredObj

Main.ThermovisorData.plot_angular_distribution_statistics — Method

plot_angular_distribution_statistics(angles,mean_D::T,std_D::T,
            lower_bound::Union{T,Nothing}=nothing,upper_bound::Union{T,Nothing}=nothing;
            length_scaler::Float64=1.0,
            label=nothing,
            minorgrid=true,
            gridlinewidth=2,
            title="Average temperature angular distribution",framestyle = :box,
            dpi=600,xlabel = L"Angle  ,°", ylabel=L"Temperature °C",
            kwargs...)      where T<:AbstractVector

The same as plot_radial_distribution_statistics but plots averaged angular distribution

Main.ThermovisorData.plot_radial_distribution_statistics — Method

plot_radial_distribution_statistics(L,mean_D::T,std_D::T,
    lower_bound::Union{T,Nothing}=nothing,upper_bound::Union{T,Nothing}=nothing;
            length_scaler::Float64=1.0,
            is_centered::Bool=true,label=nothing,
            minorgrid=true,gridlinewidth=2,title="Average temperature radial distribution",
            framestyle = :box,
            dpi=600,xlabel = "Distance  across the sample ,mm", ylabel="Temperature  °C",
            kwargs...)      where T<:AbstractVector

Plots radial ditribution averaged value, confidence bounds and confidence bounds
multiplied by the Student's coefficient

Main.ThermovisorData.points_within_line! — Method

points_within_line!(imag::AbstractMatrix,line_points::AbstractVector)

Forces all line points to lie within the possible region according toe the image size

Main.ThermovisorData.radial_distribution — Method

radial_distribution(imag::AbstractMatrix,c::CentredObj,angles_range::AbstractRange,line_length;mm_per_pixel=1.0)

Calls along_mask_line_distribution on lines oriented with some angles range and puts the resulting distribution into one matrix j'th column of this matrix corresponds to the distribution along the line oriented with ang[j] angle

Main.ThermovisorData.radial_distribution_statistics — Method

radial_distribution_statistics(along_length_coordinate,distrib;length_per_pixel=1.0,is_use_student=true)

This function evaluates mean radial diatribution, it's standard deviation and student's coefficient Input arguments along_length, distrib - distribution matrix. All rows of distrib which contains NaNs will be droped.

Optional:

maxlength - maximal value of alonglength_coordinate to be includet in to the statistics evaluation

isusestudent - flag if use students's coefficient

Main.ThermovisorData.read_temperature_file — Method

read_temperature_file(f_name::AbstractString)

Reads temeprature file f_name is a full file name

Main.ThermovisorData.rearranged_diagonal — Method

rearranged_diagonal(c::Union{SquareObj,CircleObj})

Returns diagonal points in x-y coordinates

Main.ThermovisorData.shift! — Method

shift!(c::CentredObj,x::AbstractVector)

Relative shift of centred object center

Main.ThermovisorData.std_within_mask — Method

std_within_mask(img::AbstractMatrix, c::CentredObj)

Evaluates standard deviation of temperature for all points within the CentredObj marker

Main.ThermovisorData.student_coefficient — Method

student_coefficient(degrees_of_freedom::Int, probability; digits::Int = 3, side::Int = 2)

Evaluates Student's distribution coefficient

Main.ThermovisorData.to_rgb — Method

to_rgb(image::Matrix{Float64};color_scheme::String="")

Converts matrix to rgb martix by applyting the color scheme using applycolourmap function from PerceptualColourMaps

Main.ThermovisorData.within_mask_line_points_distribution — Function

within_mask_line_points_distribution(imag::AbstractMatrix,c::CentredObj,direction_angle=0.0,line_length=10.0;use_wu::Bool=false)

Function evaluates the distribution of values in imag matrix along the line with length line_length in pixels oriented with the angle direction_angle in degrees with respect to the posistive direction of oX (column index increase), this line lies within the mask (CentreObj) and goes through its center.

Function returns:

points - vector of CartesianIndex of image's points lying on the line

distrib - distribution of values

linepoints - endpoints of line the Tupple of (leftx,leftY,rightx,right_y)